+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - TT-1577 - Move company title under logo on mobile.
 - TT-1575 - Update registration journey for Companies House listed company.
 - No ticket - Replace custom breadcrumbs with shared component
+- TT-1607 - Verification Code allow email
 
 ### Fixed bugs:
 - No ticket - Upgrade vulnerable django version to django 1.11.22